requirements.txt
gspread
gspread-dataframe
pysnooper
python-dotenv